GenAI/CustomizeResume/requirements.txt
streamlit 
langchain 
langchain-groq
langchain-community
python-docx